feat: changed the display of information about indicators in the block with information about the country on the map
