+++ src/Components/Slider/constants.js
@@ -1,9 +1,3 @@
 export const LEFT_ARROW = 'slider-left';
 export const RIGHT_ARROW = 'slider-right';
 export const MAIN_INFO_BLOCK = 'slider__info-block';
-export const DATA_TYPES_VALUES = ['TotalDeaths', 'TotalRecovered', 'TotalConfirmed',
-  'By100kDeaths', 'By100kRecovered', 'By100kConfirmed', 'NewDeathsBy100k',
-  'NewRecoveredBy100k', 'NewConfirmedBy100k', 'NewDeaths', 'NewRecovered', 'NewConfirmed'];
-export const DATA_TYPES_DECRYPTION = ['Total Deaths', 'Total Recovered', 'Total Confirmed',
-  'Total Deaths / 100 th', 'Total Recovered / 100 th', 'Total Confirmed / 100 th', 'Daily Deaths / 100 th',
-  'Daily Recovered / 100 th', 'Daily Confirmed / 100 th', 'Daily Deaths', 'Daily Recovered', 'Daily Confirmed'];
