+++ src/Constants/index.Constants.js
@@ -5,3 +5,9 @@ export const GLOBAL_SELECTOR = 'All';
 export const DEFAULT_COUNTRY = GLOBAL_SELECTOR;
 export const DEFAULT_PRECISION = 2;
 export const WORLD_POPULATION = 7594000000;
+export const DATA_TYPES_VALUES = ['TotalDeaths', 'TotalRecovered', 'TotalConfirmed',
+  'By100kDeaths', 'By100kRecovered', 'By100kConfirmed', 'NewDeathsBy100k',
+  'NewRecoveredBy100k', 'NewConfirmedBy100k', 'NewDeaths', 'NewRecovered', 'NewConfirmed'];
+export const DATA_TYPES_DECRYPTION = ['Total Deaths', 'Total Recovered', 'Total Confirmed',
+  'Total Deaths / 100 th', 'Total Recovered / 100 th', 'Total Confirmed / 100 th', 'Daily Deaths / 100 th',
+  'Daily Recovered / 100 th', 'Daily Confirmed / 100 th', 'Daily Deaths', 'Daily Recovered', 'Daily Confirmed'];
